% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind2related.R
\name{genind2related}
\alias{genind2related}
\title{Convert a genind object to 'related' format}
\usage{
genind2related(x)
}
\arguments{
\item{x}{A genind object}
}
\value{
List object used as input to 'related'
}
\description{
Convert a genind object to 'related' format
}
