% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allele_freqs.R
\name{get_allele_freqs}
\alias{get_allele_freqs}
\title{Get allele frequencies from a genind object}
\usage{
get_allele_freqs(x)
}
\arguments{
\item{x}{A genind object}
}
\value{
A tidy tibble of allele frequencies
}
\description{
Get allele frequencies from a genind object
}
