% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ind_stats.R
\name{get_ind_stats}
\alias{get_ind_stats}
\title{Calculate summary statistics by individual using a genind object}
\usage{
get_ind_stats(x)
}
\arguments{
\item{x}{A genind object}
}
\value{
A tibble of summary stats for individuals
}
\description{
Calculate summary statistics by individual using a genind object
}
