% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_diversity.R
\name{get_locus_diversity}
\alias{get_locus_diversity}
\title{Generate a tidy data frame of diversity stats for each locus}
\usage{
get_locus_diversity(x)
}
\arguments{
\item{x}{A genind object}
}
\value{
A tibble

Columns are:
n_alleles: the number of observed alleles in the population
shannon: shannon diversity (calculated from poppr)
hexp_nei: Nei's gene diversity
evenness: a measure of allelic evenness, with one being completely even representation (calculated from poppr)
}
\description{
Generate a tidy data frame of diversity stats for each locus
}
