% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_stats.R
\name{get_locus_stats}
\alias{get_locus_stats}
\title{Calculate summary statistics by locus for each population in a genind object}
\usage{
get_locus_stats(x, hwe_test = FALSE, n_boot = 1000)
}
\arguments{
\item{x}{A genind object}

\item{hwe_test}{Calculate p-value for HWE}

\item{n_boot}{Number of bootstrap replicates for HWE test}
}
\value{
A tibble with summary statistics (by locus and population)
}
\description{
Calculate summary statistics by locus for each population in a genind object
}
