% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_test.R
\name{hwe_test}
\alias{hwe_test}
\title{Test for Hardy-Weinberg equilibrium}
\usage{
hwe_test(x, n_boot = 1000)
}
\arguments{
\item{x}{A genind object}

\item{n_boot}{The number of bootstrap replicates to run}
}
\value{
A tibble with the HWE test results for each population separately
}
\description{
Test for Hardy-Weinberg equilibrium
}
