% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_fst.R
\name{pairwise_fst}
\alias{pairwise_fst}
\title{Calculate Weir and Cockerham's Fst between pairs of populations}
\usage{
pairwise_fst(x, n_bootstraps = 1000)
}
\arguments{
\item{x}{A genind object}

\item{n_bootstraps}{Number of bootstraps to calculate confidence intervals}
}
\value{
A dataframe with pairwise FST, lower and upper 95\% confidence limits
}
\description{
Calculate Weir and Cockerham's Fst between pairs of populations
}
