% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpca.R
\name{qpca}
\alias{qpca}
\title{Run a quick and dirty PCA using a genind object}
\usage{
qpca(x, n_axes = 4)
}
\arguments{
\item{x}{A genind oject}

\item{n_axes}{The number of axes to keep}
}
\value{
A tidy data frame with individual loadings
}
\description{
Run a quick and dirty PCA using a genind object
}
