% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related2matrix.R
\name{related2matrix}
\alias{related2matrix}
\title{Convert output from related 'coancestry' fuction into a matrix}
\usage{
related2matrix(
  x,
  method = c("ritland", "wang", "lynchli", "lynchrd", "quellergt", "dyadml", "trioml"),
  rows = NULL,
  cols = NULL
)
}
\arguments{
\item{x}{Output from related's coancestry() function}

\item{method}{Relatedness estimator to use}

\item{rows}{A character vector to subset the rows of the matrix}

\item{cols}{A character vector to subset the columns of the matrix}
}
\value{
A matrix of relatedness estimates
}
\description{
Convert output from related 'coancestry' fuction into a matrix
}
